refactor HmilySQLComputeEngine
